Move embedding utils in utils directory
