+++ src/omop_emb/backends/base.py
@@ -14,7 +14,7 @@ from omop_alchemy.cdm.model.vocabulary import Concept
 from .registry import ModelRegistry, ensure_model_registry_schema
 from ..config import BackendType, IndexType, MetricType
 from .errors import ModelRegistrationConflictError
-from .embedding_utils import (
+from omop_emb.utils.embedding_utils import (
     EmbeddingModelRecord, 
     EmbeddingConceptFilter, 
     NearestConceptMatch,
